fix: ErrorAggregation count 필드 타입 변경

+++ src/main/java/S13P31A306/loglens/domain/dashboard/dto/opensearch/ErrorAggregation.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public record ErrorAggregation(
         String exceptionType,
         String message,
-        Long count,
+        Integer count,
         LocalDateTime firstOccurrence,
         LocalDateTime lastOccurrence,
         String stackTrace,
